# contracts/nl2agent-openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Nl2AgentApplyLocalResourcesRequest": {
        "description": "Request body for bulk-binding local tools and skills to a draft agent.",
        "properties": {
          "recommendation_batch_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Recommendation Batch Id",
            "type": "string"
          },
          "skill_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Skill Ids",
            "type": "array"
          },
          "tool_config_values": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tool Config Values",
            "type": "object"
          },
          "tool_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Tool Ids",
            "type": "array"
          }
        },
        "required": [
          "recommendation_batch_id"
        ],
        "title": "Nl2AgentApplyLocalResourcesRequest",
        "type": "object"
      },
...
      "Nl2AgentIdentityRequest": {
        "description": "Persist the user-confirmed display name for an NL2AGENT draft.",
        "properties": {
          "display_name": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "Nl2AgentIdentityRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Runtime APIs",
    "title": "Nexent Runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/nl2agent/session/start": {
      "post": {
        "description": "Start a new NL2AGENT session: creates a draft agent and a conversation.\n\nReturns ``{\"agent_id\": int, \"conversation_id\": int, \"draft_name\": str}``.\nThe frontend then opens the chat page with the NL2AGENT default agent_id\nand this conversation_id.",
        "operationId": "start_session_api_nl2agent_session_start_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session Api"
      }
    },
    "/nl2agent/session/{agent_id}/apply-local-resources": {
      "post": {
        "description": "Bulk-bind local tools and skills to the draft agent.",
        "operationId": "apply_local_resources_api_nl2agent_session__agent_id__apply_local_resources_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentApplyLocalResourcesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Local Resources Api"
      }
    },
...
    "/nl2agent/session/{agent_id}/finalize": {
      "post": {
        "description": "Finalize the draft agent by generating its full prompt set.",
        "operationId": "finalize_agent_api_nl2agent_session__agent_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentFinalizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Agent Api"
      }
    },
    "/nl2agent/session/{agent_id}/identity": {
      "put": {
        "operationId": "save_agent_identity_api_nl2agent_session__agent_id__identity_put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Agent Identity Api"
      }
    },
    "/nl2agent/session/{agent_id}/install-web-skill": {
      "post": {
        "description": "Install a single official/web skill into the tenant.",
        "operationId": "install_web_skill_api_nl2agent_session__agent_id__install_web_skill_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentInstallWebSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Web Skill Api"
      }
    },
    "/nl2agent/session/{agent_id}/local-resources/register": {
      "post": {
        "operationId": "register_local_resources_api_nl2agent_session__agent_id__local_resources_register_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentRecommendationBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Local Resources Api"
      }
    },
    "/nl2agent/session/{agent_id}/local-resources/skip": {
      "post": {
        "operationId": "skip_local_resources_api_nl2agent_session__agent_id__local_resources_skip_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentRecommendationSkipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skip Local Resources Api"
      }
    },
    "/nl2agent/session/{agent_id}/mcp/install": {
      "post": {
        "operationId": "install_recommended_mcp_api_nl2agent_session__agent_id__mcp_install_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentMcpInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Recommended Mcp Api"
      }
    },
    "/nl2agent/session/{agent_id}/mcp/{mcp_id}/bind-tools": {
      "post": {
        "operationId": "bind_mcp_tools_api_nl2agent_session__agent_id__mcp__mcp_id__bind_tools_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "mcp_id",
            "required": true,
            "schema": {
              "title": "Mcp Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentMcpBindToolsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bind Mcp Tools Api"
      }
    },
    "/nl2agent/session/{agent_id}/mcp/{mcp_id}/skip-tools": {
      "post": {
        "operationId": "skip_mcp_tools_api_nl2agent_session__agent_id__mcp__mcp_id__skip_tools_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "mcp_id",
            "required": true,
            "schema": {
              "title": "Mcp Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skip Mcp Tools Api"
      }
    },
    "/nl2agent/session/{agent_id}/models": {
      "put": {
        "operationId": "select_models_api_nl2agent_session__agent_id__models_put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Nl2AgentModelSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Models Api"
      }
    },
...
  }
}